Migrated to Django 4.2.1.
Migrated to Django 4.2.1 and python 3.8 or 3.9 (3.10+ has an issue with building backports.zoneinfo which is a whitenoise dependency). Fixed any errors and warning outlined by the checker. Migrarted to as many latest versions of dependencies as it was possible. Applied migrations and added those to the shipped database. Removed the __pycache__ from the repo and prevented those files from being added in the future. NB! All the function ality has not been thoroughly tested yet and there could be potential issues present (example: solical login).